Remove mappings from MailMapper
Nextras\Orm 2.1.0 shows an error when adding a mapping to a reflection
and mapping for the key already exists. These mappings were introduced
in the past because the library required them or something. Now it does
not need them so we can just remove them.

See: https://github.com/nextras/orm/issues/174